Merged dev into nightly

+++ src/game/geo_misc.h
@@ -6,9 +6,9 @@
 
 enum FlyingCarpetState
 {
-    FLYING_CARPET_IDLE = 0,
+    FLYING_CARPET_IDLE                 = 0,
     FLYING_CARPET_MOVING_WITHOUT_MARIO = 1,
-    FLYING_CARPET_MOVING_WITH_MARIO = 2
+    FLYING_CARPET_MOVING_WITH_MARIO    = 2
 };
 
 extern s8 gFlyingCarpetState;
@@ -22,7 +22,7 @@ extern void make_vertex(
     Vtx *vtx, s32 n, f32 x, f32 y, f32 z, s16 tx, s16 ty, u8 r, u8 g, u8 b, u8 a
 );
 #endif
-extern s16 round_float(f32);
+extern s32 round_float(f32);
 extern Gfx *geo_exec_inside_castle_light(s32 callContext, struct GraphNode *node, UNUSED Mat4 mtx);
 extern Gfx *geo_exec_flying_carpet_timer_update(s32 callContext, struct GraphNode *node, UNUSED Mat4 mtx);
 extern Gfx *geo_exec_flying_carpet_create(s32 callContext, struct GraphNode *node, UNUSED Mat4 mtx);
