+++ src/game/ingame_menu.h
@@ -63,52 +63,52 @@ enum SpecialFontChars {
 // definitions for some of the special characters defined in charmap.txt
 enum DialogSpecialChars {
 #ifdef VERSION_EU
-    DIALOG_CHAR_LOWER_A_GRAVE = 0x60,      // 'a' grave
+    DIALOG_CHAR_LOWER_A_GRAVE      = 0x60, // 'a' grave
     DIALOG_CHAR_LOWER_A_CIRCUMFLEX = 0x61, // 'a' circumflex
-    DIALOG_CHAR_LOWER_A_UMLAUT = 0x62,     // 'a' umlaut
-    DIALOG_CHAR_UPPER_A_GRAVE = 0x64,      // 'A' grave
+    DIALOG_CHAR_LOWER_A_UMLAUT     = 0x62, // 'a' umlaut
+    DIALOG_CHAR_UPPER_A_GRAVE      = 0x64, // 'A' grave
     DIALOG_CHAR_UPPER_A_CIRCUMFLEX = 0x65, // 'A' circumflex
-    DIALOG_CHAR_UPPER_A_UMLAUT = 0x66,     // 'A' umlaut
-    DIALOG_CHAR_LOWER_E_GRAVE = 0x70,      // 'e' grave
+    DIALOG_CHAR_UPPER_A_UMLAUT     = 0x66, // 'A' umlaut
+    DIALOG_CHAR_LOWER_E_GRAVE      = 0x70, // 'e' grave
     DIALOG_CHAR_LOWER_E_CIRCUMFLEX = 0x71, // 'e' circumflex
-    DIALOG_CHAR_LOWER_E_UMLAUT = 0x72,     // 'e' umlaut
-    DIALOG_CHAR_LOWER_E_ACUTE = 0x73,      // 'e' acute
-    DIALOG_CHAR_UPPER_E_GRAVE = 0x74,      // 'E' grave
+    DIALOG_CHAR_LOWER_E_UMLAUT     = 0x72, // 'e' umlaut
+    DIALOG_CHAR_LOWER_E_ACUTE      = 0x73, // 'e' acute
+    DIALOG_CHAR_UPPER_E_GRAVE      = 0x74, // 'E' grave
     DIALOG_CHAR_UPPER_E_CIRCUMFLEX = 0x75, // 'E' circumflex
-    DIALOG_CHAR_UPPER_E_UMLAUT = 0x76,     // 'E' umlaut
-    DIALOG_CHAR_UPPER_E_ACUTE = 0x77,      // 'E' acute
-    DIALOG_CHAR_LOWER_U_GRAVE = 0x80,      // 'u' grave
+    DIALOG_CHAR_UPPER_E_UMLAUT     = 0x76, // 'E' umlaut
+    DIALOG_CHAR_UPPER_E_ACUTE      = 0x77, // 'E' acute
+    DIALOG_CHAR_LOWER_U_GRAVE      = 0x80, // 'u' grave
     DIALOG_CHAR_LOWER_U_CIRCUMFLEX = 0x81, // 'u' circumflex
-    DIALOG_CHAR_LOWER_U_UMLAUT = 0x82,     // 'u' umlaut
-    DIALOG_CHAR_UPPER_U_GRAVE = 0x84,      // 'U' grave
+    DIALOG_CHAR_LOWER_U_UMLAUT     = 0x82, // 'u' umlaut
+    DIALOG_CHAR_UPPER_U_GRAVE      = 0x84, // 'U' grave
     DIALOG_CHAR_UPPER_U_CIRCUMFLEX = 0x85, // 'U' circumflex
-    DIALOG_CHAR_UPPER_U_UMLAUT = 0x86,     // 'U' umlaut
+    DIALOG_CHAR_UPPER_U_UMLAUT     = 0x86, // 'U' umlaut
     DIALOG_CHAR_LOWER_O_CIRCUMFLEX = 0x91, // 'o' circumflex
-    DIALOG_CHAR_LOWER_O_UMLAUT = 0x92,     // 'o' umlaut
+    DIALOG_CHAR_LOWER_O_UMLAUT     = 0x92, // 'o' umlaut
     DIALOG_CHAR_UPPER_O_CIRCUMFLEX = 0x95, // 'O' circumflex
-    DIALOG_CHAR_UPPER_O_UMLAUT = 0x96,     // 'O' umlaut
+    DIALOG_CHAR_UPPER_O_UMLAUT     = 0x96, // 'O' umlaut
     DIALOG_CHAR_LOWER_I_CIRCUMFLEX = 0xA1, // 'i' circumflex
-    DIALOG_CHAR_LOWER_I_UMLAUT = 0xA2,     // 'i' umlaut
-    DIALOG_CHAR_I_NO_DIA = 0xEB,           // 'i' without diacritic
-    DIALOG_CHAR_DOUBLE_LOW_QUOTE = 0xF0,   // German opening quotation mark
+    DIALOG_CHAR_LOWER_I_UMLAUT     = 0xA2, // 'i' umlaut
+    DIALOG_CHAR_I_NO_DIA           = 0xEB, // 'i' without diacritic
+    DIALOG_CHAR_DOUBLE_LOW_QUOTE   = 0xF0, // German opening quotation mark
 #endif
 #if defined(VERSION_US) || defined(VERSION_EU)
-    DIALOG_CHAR_SLASH = 0xD0,
-    DIALOG_CHAR_MULTI_THE = 0xD1, // 'the'
-    DIALOG_CHAR_MULTI_YOU = 0xD2, // 'you'
+    DIALOG_CHAR_SLASH                = 0xD0,
+    DIALOG_CHAR_MULTI_THE            = 0xD1, // 'the'
+    DIALOG_CHAR_MULTI_YOU            = 0xD2, // 'you'
 #endif
-    DIALOG_CHAR_PERIOD = 0x6E,
-    DIALOG_CHAR_COMMA = 0x6F,
-    DIALOG_CHAR_COLOR = 0xDF,
-    DIALOG_CHAR_SPACE = 0x9E,
-    DIALOG_CHAR_STAR_COUNT = 0xE0, // number of stars
-    DIALOG_CHAR_UMLAUT = 0xE9,
-    DIALOG_CHAR_DAKUTEN = 0xF0,
+    DIALOG_CHAR_PERIOD               = 0x6E,
+    DIALOG_CHAR_COMMA                = 0x6F,
+    DIALOG_CHAR_COLOR                = 0xDF,
+    DIALOG_CHAR_SPACE                = 0x9E,
+    DIALOG_CHAR_STAR_COUNT           = 0xE0, // number of stars
+    DIALOG_CHAR_UMLAUT               = 0xE9,
+    DIALOG_CHAR_DAKUTEN              = 0xF0,
     DIALOG_CHAR_PERIOD_OR_HANDAKUTEN = 0xF1,
-    DIALOG_CHAR_STAR_FILLED = 0xFA,
-    DIALOG_CHAR_STAR_OPEN = 0xFD,
-    DIALOG_CHAR_NEWLINE = 0xFE,
-    DIALOG_CHAR_TERMINATOR = 0xFF
+    DIALOG_CHAR_STAR_FILLED          = 0xFA,
+    DIALOG_CHAR_STAR_OPEN            = 0xFD,
+    DIALOG_CHAR_NEWLINE              = 0xFE,
+    DIALOG_CHAR_TERMINATOR           = 0xFF
 };
 
 // gDialogResponse
@@ -137,17 +137,17 @@ void print_hud_lut_string(s8 hudLUT, s16 x, s16 y, const u8 *str);
 void print_menu_generic_string(s16 x, s16 y, const u8 *str);
 void handle_menu_scrolling(s8 scrollDirection, s8 *currentIndex, s8 minIndex, s8 maxIndex);
 #if defined(VERSION_US) || defined(VERSION_EU)
-s16 get_str_x_pos_from_center(s16 centerPos, u8 *str, f32 scale);
+s32 get_str_x_pos_from_center(s16 centerPos, u8 *str, f32 scale);
 #endif
 #if defined(VERSION_JP) || defined(VERSION_SH)
 #define get_str_x_pos_from_center get_str_x_pos_from_center_scale
 #endif
 #if defined(VERSION_JP) || defined(VERSION_EU) || defined(VERSION_SH)
-s16 get_str_x_pos_from_center_scale(s16 centerPos, u8 *str, f32 scale);
+s32 get_str_x_pos_from_center_scale(s16 centerPos, u8 *str, f32 scale);
 #endif
 void print_hud_my_score_coins(s32 useCourseCoinScore, s8 fileNum, s8 courseNum, s16 x, s16 y);
 void int_to_str(s32 num, u8 *dst);
-s16 get_dialog_id(void);
+s32 get_dialog_id(void);
 void create_dialog_box(s16 dialog);
 void create_dialog_box_with_var(s16 dialog, s32 dialogVar);
 void create_dialog_inverted_box(s16 dialog);
@@ -162,6 +162,6 @@ void set_cutscene_message(s16 xOffset, s16 yOffset, s16 msgIndex, s16 msgDuratio
 void do_cutscene_handler(void);
 void render_hud_cannon_reticle(void);
 void reset_red_coins_collected(void);
-s16 render_menus_and_dialogs(void);
+s32 render_menus_and_dialogs(void);
 
 #endif // INGAME_MENU_H
